move test to avoid introducing additional dependencies to migration_assistant
